Another shot at ruby_version_is and ruby_bug guards.
The guard takes either a String or a Range to specify
which version(s) match. For example, to run the specs on
any version >= 1.8.6 p114

  ruby_version_is "1.8.6.114" do ... end

To run the specs on any version from 1.8.6 to 1.8.6 p114

  ruby_version_is "1.8.6".."1.8.6.114" do ... end

To run the specs on any version from 1.8.6 to 1.8.7 but
*not* including 1.8.7

  ruby_version_is "1.8.6"..."1.8.7" do ... end

Also, ruby_bug guard now takes a version 2nd argument. The
ruby_bug guard will not run the specs on any version of MatzRuby
that is <= 'version'. For example,

  ruby_bug "#2", "1.8.6.114" do ... end

will not run the specs on MatzRuby for any version less than
1.8.6 p114, but it will run the specs on any version after
that and on any other implementation besides MatzRuby and
regardless of version.