Added an object in which to instance_eval each spec file.
This fixes some issues with e.g. IO.popen specs. This still does
not follow RSpec completely. They use a pristine object for each
example. Adding all this to MSpec probably tips over the complexity
threshold. We can revisit if it appears to be essential, but for
now, this provides enough "feature" compatibility for any specs
that we have as far as I'm aware.