Replaced tags dir with tags pattern. See below.
Previously, a tag file for a particular spec file was
generated by finding a "stem" of the spec file and replacing
that stem with the tags directory.

This has been replaced with an arbitrary number of
substitution patterns (i.e. [Regexp, String]) that are
applied to the spec filename to generate the tag filename.

In the 'default.mspec' or other config file, add an entry
with the following structure:

  set :tags_patterns, [[/a/, "b"], [/c/, "d"]]

If no entry is set, the default transformation is:

  [[%r(/spec/), '/spec/tags/'], [/_spec.rb$/, '_tags.txt']]