Added back loading 'default.mspec' first.
The scripts will first attempt to load 'default.mspec'. If that fails,
they will then look for a config file based on RUBY_ENGINE and RUBY_VERSION
as detailed previously. This is necessary because the default 'ruby'
executable on PATH is used to run the mspec front script by default.
If that is MRI, it will load e.g. 'ruby.1.8.mspec' and never load
the implementation-specific config file.