Added config[:prefix] and appending _spec.rb to file patterns.
The config[:prefix] value will be prepended to file patterns to locate
files and directories to run. Additionally, _spec.rb will be appended
to check for a file match. The following examples demonstrate:

  mspec spec/1.8/core/array/append_spec.rb
  mspec core/array/append

In the first, the full path to the file is given. The second example
is equivalent if config[:prefix] = 'spec/1.8'.