Added -C, --chdir DIR option.
This option will change the working directory via Dir.chdir at
the point the option is processed. For example, suppose the
working directory is 'files' and 'files/rubyspec' is the root
of the rubyspec files. The command

  mspec -C rubyspec -B ruby.1.8.mspec

will run all the specs listed in the :files config option. The
files will be found because the working directory will be
'files/rubyspec' and the :files entries are relative to that
directory.

Note that in the example above, the explicit -B option is necessary
because the command line options override the default config options
ond the default config file is loaded before command line options
are processed.