Added ability for mspec :language and other collections.
If mspec is passed a :some_string type argument, it will attempt
to look up a config value, which can be a single value or an array,
and attempt to process it for spec files to run. This simplifies
running e.g. all the library specs now that the ones for 1.8 and
1.9 are colocated.