Removed Gemfile.lock for mspec.
In RSpec, raise_error and raise_exception are aliases. Since "error"
is a more general term, it's reasonable that it matches "kinds of"
exceptions. Here, raise_exception is used to match only a specific
kind of exception. Any confusion that results will be short-lived
as MSpec 2.0 will use a different syntax.